+++ .storybook/config.js
@@ -1,7 +1,7 @@
 import { configure } from '@storybook/react';
 
 // automatically import all files ending in *.stories.js
-const req = require.context('../src', true, /(.*)\.stories\.jsx?$/);
+const req = require.context('../frog/imports/client', true, /\.stories\.jsx?$/);
 function loadStories() {
   req.keys().forEach(filename => req(filename));
 }
