--- frog-ui/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "frog-ui",
-  "version": "1.0.0",
-  "private": true,
-  "main": "src/index.js",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "react": "^16.8.6"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.5.4",
-    "@storybook/addon-actions": "^5.1.9",
-    "@storybook/addon-links": "^5.1.9",
-    "@storybook/addons": "^5.1.9",
-    "@storybook/react": "^5.1.9",
-    "babel-loader": "^8.0.6"
-  },
-  "scripts": {
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  }
-}
