--- frog-ui/src/index.js
@@ -1,3 +1,0 @@
-// @flow
-
-export { Button } from './Button';
