+++ frog/package.json
@@ -38,7 +38,6 @@
     "file-dialog": "^0.0.7",
     "file-saver": "^2.0.0",
     "font-awesome-filetypes": "^2.0.1",
-    "frog-ui": "^1.0.0",
     "html2canvas": "^1.0.0-alpha.12",
     "http": "0.0.0",
     "import-all.macro": "^2.0.3",
