+++ package.json
@@ -15,7 +15,6 @@
   "workspaces": [
     "ac/*",
     "op/*",
-    "frog-ui",
     "frog",
     "templates/activity",
     "templates/operator"
@@ -36,11 +35,16 @@
     "@babel/preset-env": "7.3.1",
     "@babel/preset-react": "7.0.0",
     "@babel/runtime": "7.3.1",
+    "@storybook/addon-actions": "^5.1.9",
+    "@storybook/addon-links": "^5.1.9",
+    "@storybook/addons": "^5.1.9",
+    "@storybook/react": "^5.1.9",
     "@testing-library/cypress": "^4.0.4",
     "assert": "^1.4.1",
     "babel-bridge": "^1.12.11",
     "babel-core": "^7.0.0-0",
     "babel-eslint": "10.0.1",
+    "babel-loader": "^8.0.6",
     "babel-plugin-captains-log": "^1.1.0",
     "chokidar": "^2.1.0",
     "cypress": "^3.3.1",
